+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 -   support for cloudwatch probes and actions
 -   support for invoking lambdas
 -   support for getting and updating lambda timeout and memory size limits
+-   probe for cloudwatch metric statistics
 
 ## [0.8.0][]